add order and fix client requirements

+++ app/Modules/General/Controllers/GeneralController.php
@@ -34,9 +34,9 @@ class GeneralController extends Controller
     {
         $country = Country::find($id);
         if ($country) {
-            $cities = City::where('country_id', $country->id)->with('zipcodes','zipcodes.city')->get();
+            $cities = City::where('country_id', $country->id)->with('zipcodes', 'zipcodes.city')->get();
 
-            return view('General::showUpdateCountry', compact('country','cities'));
+            return view('General::showUpdateCountry', compact('country', 'cities'));
         }
         return view('General::notFound');
 
@@ -45,9 +45,27 @@ class GeneralController extends Controller
     public function handleDeleteCountry($id)
     {
         $country = Country::find($id);
+
         if ($country) {
+
+            if ($country->companies->count() > 0) {
+                alert()->error('Oups!', 'Le pays est déja affecté à ' . $country->companies->count() . ' societé(s) ! ')->persistent('Fermer');
+                return redirect()->back();
+
+            }
+            if ($country->stores->count() > 0) {
+                alert()->error('Oups!', 'Le pays est déja affecté à ' . $country->stores->count() . ' magasin(s) ! ')->persistent('Fermer');
+                return redirect()->back();
+
+            }
+
+            if ($country->warehouses->count() > 0) {
+                alert()->error('Oups!', 'Le pays est déja affecté à ' . $country->warehouses->count() . 'entrepôt(s) ! ')->persistent('Fermer');
+                return redirect()->back();
+
+            }
             $country->delete();
-            alert()->success('Succès!','Le pays a été supprimé avec succès')->persistent('Fermer');
+            alert()->success('Succès!', 'Le pays a été supprimé avec succès')->persistent('Fermer');
             return redirect()->back();
         }
         alert()->error('Pays introuvable')->persistent('Fermer');
